Added Automation
Added some exceptions and correct value statements throughout the program to add more user-friendliness to the program. Also to remove some crashes that cause the program to fail if wrong values are entered or after the output file is saved and unable to be used due to wrong format (hopefully).

Also added a lot of automation for the user.