Added a Residuals Plot
I have added  in a way to plot the residuals of a certain polynomial degree fit. So if the user needs to be able to see the original data with the model fit and the residuals at the same time, they are plotted in the same figure to be looked at.