Create reading_in_stats.py
Reads in an excel spreadsheet from a sample active report. The code then counts the unique items in the list of item names and adds up the total amount of times that unique item appears and creates a dictionary with keys (item name) and values (total amount). The program then outputs those keys and values to a .txt file for future use.